services/frontend/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
SQLAlchemy==1.4.44
Werkzeug==2.2.2
zipp==3.11.0
minio == 7.1.12
